+++ src/libsac2c/tree/tree_compound.h
@@ -636,7 +636,7 @@ extern node *TCappendAssign (node *assign_chain, node *assign);
 
 extern node *TCappendAssignIcm (node *assign, char *name, node *args);
 
-extern int TCcountAssigns (node *assigns);
+extern size_t TCcountAssigns (node *assigns);
 extern node *TCgetLastAssign (node *arg_node);
 
 /*--------------------------------------------------------------------------*/
